+++ src/schemathesis/cli/commands/run/__init__.py
@@ -46,6 +46,7 @@ from schemathesis.cli.options import (
     PROXY,
     RATE_LIMIT,
     REPORT,
+    REPORT_ALLURE_PATH,
     REPORT_DIR,
     REPORT_HAR_PATH,
     REPORT_JUNIT_PATH,
@@ -131,6 +132,7 @@ DEFAULT_PHASES = ["examples", "coverage", "fuzzing", "stateful"]
 @grouped_option(*REPORT_VCR_PATH.args, **REPORT_VCR_PATH.kwargs)
 @grouped_option(*REPORT_HAR_PATH.args, **REPORT_HAR_PATH.kwargs)
 @grouped_option(*REPORT_NDJSON_PATH.args, **REPORT_NDJSON_PATH.kwargs)
+@grouped_option(*REPORT_ALLURE_PATH.args, **REPORT_ALLURE_PATH.kwargs)
 @grouped_option(*REPORT_PRESERVE_BYTES.args, **REPORT_PRESERVE_BYTES.kwargs)
 @grouped_option(*OUTPUT_SANITIZE.args, **OUTPUT_SANITIZE.kwargs)
 @grouped_option(*OUTPUT_TRUNCATE.args, **OUTPUT_TRUNCATE.kwargs)
@@ -211,6 +213,7 @@ def run(
     report_vcr_path: LazyFile | None = None,
     report_har_path: LazyFile | None = None,
     report_ndjson_path: LazyFile | None = None,
+    report_allure_path: str | None = None,
     report_preserve_bytes: bool | None = None,
     output_sanitize: bool | None = None,
     output_truncate: bool | None = None,
@@ -271,6 +274,7 @@ def run(
         vcr_path=report_vcr_path.name if report_vcr_path else None,
         har_path=report_har_path.name if report_har_path else None,
         ndjson_path=report_ndjson_path.name if report_ndjson_path else None,
+        allure_path=report_allure_path,
         directory=Path(report_directory),
         preserve_bytes=report_preserve_bytes,
     )