+++ src/schemathesis/reporting/__init__.py
@@ -4,4 +4,10 @@ from schemathesis.reporting.junitxml import JunitXmlWriter
 from schemathesis.reporting.ndjson import NdjsonWriter
 from schemathesis.reporting.vcr import VcrWriter
 
-__all__ = ["HarWriter", "JunitXmlWriter", "NdjsonWriter", "VcrWriter", "get_command_representation"]
+__all__ = [
+    "HarWriter",
+    "JunitXmlWriter",
+    "NdjsonWriter",
+    "VcrWriter",
+    "get_command_representation",
+]